requirements.txt
certifi==2019.6.16
paho-mqtt==1.4.0
